fix to schemas and models

+++ flask_app/blueprints/restapi/ns_recipe.py
@@ -961,8 +961,6 @@ class RecipeListResource(Resource):
 
         json_data = request.get_json()
 
-        args = parser.parse_args()
-
         ## verify user
 
         user_id = get_jwt_identity()
@@ -979,27 +977,8 @@ class RecipeListResource(Resource):
             log.error("User couln't be found.")
             return Response(status=400, response="User couln't be found.")
 
-        if user.user_type != USER_TYPE.ADMIN.value:
-            return Response(status=403)
-
-        ## verify company
-
-
-        try:
-            user = UserDB.get(args['id'])
-        except peewee.DoesNotExist:
-            # Otherwise block user token (user cant be logged in and stil reach this far)
-            # this only occurs when accounts are not in db, so in prod this wont happen
-            jti = get_jwt()["jti"]
-            now = datetime.now(timezone.utc)
-            token_block_record = TokenBlocklist(jti=jti, created_at=now)
-            token_block_record.save()
-            log.error("User couln't be found.")
-            return Response(status=400, response="User couln't be found.")
-
-        print(user.user_type in USER_TYPE_SET)
         if user.user_type != USER_TYPE.COMPANY.value:
-            return Response(status=400, response="User isn't a company.")
+            return Response(status=403,response="User is not a company.")
 
         # Validate args by loading it into schema
 
@@ -1010,7 +989,7 @@ class RecipeListResource(Resource):
 
         # Change get or create needed objects
         # removing because the must be transformed before entity building
-        nutrition_table = recipe_validated.pop('nutrition_informations')
+
         preparation = recipe_validated.pop('preparation')
         ingredients = recipe_validated.pop('ingredients')
         tags = recipe_validated.pop('tags')
@@ -1019,21 +998,27 @@ class RecipeListResource(Resource):
         recipe = RecipeDB(**recipe_validated)
         recipe.preparation = str(preparation).encode()
         # use .decode() to decode
-        recipe.save()
+        recipe.created_by = user
+        
 
         # build relation to nutrition_table
+        if 'nutrional_table' in recipe_validated:
+            nutrition_table = recipe_validated.pop('nutrional_table')
+            try:
+                if 'id' in nutrition_table:
+                    nutrition_table.pop('id')
 
-        try:
-            if 'id' in nutrition_table:
-                nutrition_table.pop('id')
+                nutrition_information = NutritionInformationDB(**nutrition_table)
+                nutrition_information.save()
+                recipe.nutrition_information = nutrition_information
+            except Exception as e:
+                nutrition_information.delete()
+                recipe.delete_instance(recursive=True)
+                return Response(status=400, response="Nutrition Table has some error.\n" + str(e))
 
-            nutrition_information = NutritionInformationDB(**nutrition_table)
-            nutrition_information.recipe = recipe
-            nutrition_information.save()
-            recipe.nutrition_informations = nutrition_information
-        except Exception as e:
-            recipe.delete_instance(recursive=True)
-            return Response(status=400, response="Nutrition Table has some error.\n" + str(e))
+        recipe.save()
+        ## recipe needs to be saved after foreign key's but before multiple to multiple relations
+        # because to build these last one recipe needs to already have an id, wich is done by save()
 
         # build multi to multi relation to tags
 
@@ -1046,19 +1031,18 @@ class RecipeListResource(Resource):
 
 
         except Exception as e:
+
             recipe.delete_instance(recursive=True)
             return Response(status=400, response="Tags Table has some error.\n" + str(e))
 
-        ## recipe needs to be saved after foreign key's but before multiple to multiple relations
-        # because to build these last one recipe needs to already have an id, wich is done by save()
-        recipe.save()
+
 
         # build multi to multi relation to Ingredient Quantity
 
         try:
             if ingredients and ingredients != {}:
                 for i in ingredients:
-                    ingredient, created = IngredientQuantity.get_or_create(name=i['name'])
+                    ingredient, created = Ingredient.get_or_create(name=i['ingredient']['name'])
                     if created:
                         ingredient.save()
                     if i['quantity_original']:
@@ -1067,32 +1051,17 @@ class RecipeListResource(Resource):
                         except Exception as e:
                             quantity_normalized = float(0)
                             log.error("Tags Table has some error...")
-                    ingredient_quantity = Ingredient(quantity_original=i['quantity_original'],
-                                                     quantity_normalized=quantity_normalized)
+                    ingredient_quantity = IngredientQuantity(quantity_original=i['quantity_original'],
+                                                             quantity_normalized=quantity_normalized)
                     ingredient_quantity.ingredient = ingredient
                     ingredient_quantity.recipe = recipe
                     ingredient_quantity.save()
-
         except Exception as e:
             recipe.delete_instance(recursive=True)
             log.error("Tags Table has some error...")
             return Response(status=400, response="Ingredients Table has some error.\n" + str(e))
 
-        # build relation to recipe_background
-
-        try:
-            recipe_background = RecipeBackgroundDB()
-            recipe_background.user = user
-            recipe_background.recipe = recipe
-            recipe_background.type = "CREATED"
-            recipe_background.save()
-        except Exception as e:
-            recipe.delete_instance(recursive=True)
-            log.error("Recipe Background Table has some error...")
-            return Response(status=400, response="Recipe Background Table has some error.\n" + str(e))
-
-        # finally build full object
-
         recipe.save()
+
         log.info("Finished POST /recipe/list")
         return Response(status=201)
